Add system tray, and some notifications as popups
