+++ battery_advisor/__init__.py
@@ -1,3 +1,2 @@
-from .utils import get_battery_status, notify, notify_with_actions, execute_action
-from .settings_loader import load_settings
 from .entry import main
+from .settings_loader import load_settings
