+++ battery_advisor/battery_advisor.py
@@ -1,11 +1,13 @@
-import time
-from .utils import get_battery_status, notify, notify_with_actions, execute_action
-from .settings_loader import load_settings
-from .gui.alerts import MessageAlert, AlertWithButtons
-from .tray import get_icon
-
 import threading
+import time
+
 from pystray import Menu, MenuItem
+
+from .gui.alerts import AlertWithButtons, MessageAlert
+from .settings_loader import load_settings
+from .tray import get_icon
+from .utils import execute_action, get_battery_status
+from .notifications import notify, notify_with_actions, alert_with_options, alert
 
 settings = load_settings()
 
@@ -14,7 +16,8 @@ class BatteryAdvisor:
     LOW_BATTERY_TRESHOLD = settings["tresholds"]["low_battery_treshold"]
     CRITICAL_BATTERY_TRESHOLD = settings["tresholds"]["critical_battery_treshold"]
     BATTERY_ACTION_TRESHOLD = settings["tresholds"]["battery_action_treshold"]
-    CHECK_INTERVAL = settings["advisor"]["check_interval"]
+    # CHECK_INTERVAL = settings["advisor"]["check_interval"]
+    CHECK_INTERVAL = 3
 
     # Configs
     NOTIFY_PLUGGED = settings["advisor"]["notify_plugged"]
@@ -33,7 +36,7 @@ class BatteryAdvisor:
 
         while True:
             if not self.running:
-                time.sleep(1)
+                time.sleep(3)
                 continue
 
             print("Checking battery status...")
@@ -50,23 +53,20 @@ class BatteryAdvisor:
 
             if plugged:
                 print("Battery is charging. Skipping checks.")
-                print("Sleeping...")
                 time.sleep(self.CHECK_INTERVAL)
                 continue
 
             # Battery Low notifications
             if batt_percent <= self.BATTERY_ACTION_TRESHOLD:
-                notify(
-                    "Battery Action",
-                    f"Your battery is at {batt_percent}%. Your system will {settings['advisor']['battery_action'].capitalize()} in a few.",
-                )
                 print("Reporting battery action.")
-                print("Sleeping...")
-                time.sleep(5)
+                alert(
+                    f"Your battery is at {int(batt_percent)}%. Your system will {settings['advisor']['battery_action'].capitalize()} in a few.",
+                )
+
                 configured_action = settings["advisor"]["battery_action"]
                 action_cmd = settings["actions"][configured_action]
                 print("Executing Battery Action. Goodbye...")
-                execute_action(action_cmd)
+                # execute_action(action_cmd)
 
             if batt_percent <= self.CRITICAL_BATTERY_TRESHOLD:
                 print("Reporting critical battery.")
@@ -82,13 +82,12 @@ class BatteryAdvisor:
 
             elif batt_percent <= self.LOW_BATTERY_TRESHOLD:
                 print("Reporting low battery.")
-                remind_time = notify_with_actions(
-                    title="Low Battery",
-                    message=f"Consider plugging your device.",
+                remind_time = alert_with_options(
+                    message=f"Low Battery. Consider plugging your device.",
                     options=self.LOW_BATTERY_OPTIONS,
-                    actions=settings["actions"],
-                    remind_time=settings["advisor"]["remind_time"],
                 )
+
+                print(remind_time)
 
             if remind_time > 0:
                 print("A function returned a remind time!")
