+++ battery_advisor/gui/alerts.py
@@ -2,6 +2,7 @@ import gi
 
 gi.require_version("Gtk", "3.0")
 from gi.repository import Gtk
+
 from ..utils import execute_action
