+++ pyproject.toml
@@ -12,10 +12,6 @@ include = ["defaultSettings.toml", "icon.png", "LICENSE"]
 
 [tool.poetry.dependencies]
 python = "^3.11"
-psutil = "^6.0.0"
-toml = "^0.10.2"
-pystray = "^0.19.5"
-pillow = "^10.4.0"
 
 [tool.poetry.group.dev.dependencies]
 black = "^24.8.0"
